bora

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: fmilheir <fmilheir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/11 11:44:35 by fmilheir          #+#    #+#             */
-/*   Updated: 2022/05/11 16:43:10 by fmilheir         ###   ########.fr       */
+/*   Updated: 2022/05/11 17:00:33 by fmilheir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int	ft_putchar(char c, int *res);
 int	ft_putstr(char *s, int *res);
 int	ft_putdec(int n, int *res);
 int	ft_putuns(unsigned int n, int *res);
-int	ft_puthex(unsigned int n, int *res);
+int	ft_puthex(unsigned long int n, int *res);
 int	ft_puthup(unsigned int num, int *res);
 int	ft_putpnt(unsigned long int n, int *res);
 #endif
